+++ tradin/src/components/units/backtest/main/BackTest.types.ts
@@ -1,0 +1,6 @@
+export interface Trade {
+    entryTime: string;
+    exitTime: string;
+    profit: number;
+    strategy: string;
+}
